PAL-391 moved MultiPolicy and Policy to requests
