+++ palisade-service/src/test/java/uk/gov/gchq/palisade/service/palisade/service/ResultAggregationServiceTest.java
@@ -33,8 +33,8 @@ import uk.gov.gchq.palisade.resource.impl.SystemResource;
 import uk.gov.gchq.palisade.service.ConnectionDetail;
 import uk.gov.gchq.palisade.service.SimpleConnectionDetail;
 import uk.gov.gchq.palisade.service.palisade.config.ApplicationConfiguration;
-import uk.gov.gchq.palisade.service.palisade.policy.MultiPolicy;
-import uk.gov.gchq.palisade.service.palisade.policy.Policy;
+import uk.gov.gchq.palisade.service.palisade.request.MultiPolicy;
+import uk.gov.gchq.palisade.service.palisade.request.Policy;
 import uk.gov.gchq.palisade.service.palisade.repository.HashMapBackingStore;
 import uk.gov.gchq.palisade.service.palisade.repository.SimpleCacheService;
 import uk.gov.gchq.palisade.service.palisade.request.AuditRequest;
