+++ policy-service/mvn_dependency_tree.txt
@@ -2,7 +2,6 @@ uk.gov.gchq.palisade:policy-service:jar:0.4.0-SNAPSHOT
 +- com.github.ben-manes.caffeine:caffeine:jar:2.8.0:compile
 |  +- org.checkerframework:checker-qual:jar:2.10.0:compile
 |  \- com.google.errorprone:error_prone_annotations:jar:2.3.3:compile
-+- com.google.protobuf:protobuf-java:jar:3.5.1:compile
 +- com.netflix.eureka:eureka-client:jar:1.9.16:compile
 |  +- org.codehaus.jettison:jettison:jar:1.3.7:runtime
 |  |  \- stax:stax-api:jar:1.0.1:runtime
